requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
